Reworked how Rooms and corridors, now Hallways, work to give more flexibility in the future
Before, corridors were just added to the board without having an object
represent them. To fix this I created a Hallway class with two child
classes, one for vertical and one for horizontal. In doing this I
realized that I didn't want Hallways to be Rooms so I created the new
parent class, Area, of which Hallway and Room are children. This made it
easier to manage the differences between them.